Sensor update functions no longer directly update reported metric values.
This way sensor update code is detached from output formatting. For future rewrites.

As a side-effect all values are updated at once (resp. almost, except metric update)
i.e. whole iteration is reported and updated as one piece. Not step by step.
It is not a big practical difference because update was short (~4.2ms on Rpi3).